MS/requirements.txt
pandas
psycopg2
django
sqlalchemy
djangorestframework
requests
django-cors-headers